realloc_06: Structure refactor, remove legacy build toos
Move code to under src and rename core to realloc. Update project to exclusively use pyproject.toml and remove setup.py. Refactor all import statements to reflect new structure.
