+++ .github/workflows/ci.yml
@@ -30,7 +30,7 @@ jobs:
 
       - name: 🧪 Run tests and coverage
         run: |
-          pytest --cov=core --cov-report=term --cov-report=html
+          pytest --cov=realloc --cov-report=term --cov-report=html
 
       - name: 📤 Upload coverage report
         uses: actions/upload-artifact@v4
