--- setup.cfg
@@ -1,9 +1,0 @@
-[metadata]
-name=trade_generator
-version=1.0.0
-description='Generates trades from positions'
-author='Ryeland Gongora'
-author_email='rsgmon@gmail.com'
-
-[options]
-packages = find:
