+++ src/realloc/cli/partial_rebalance_main.py
@@ -1,6 +1,6 @@
 import argparse
 import json
-from core import (
+from realloc import (
     Account,
     PortfolioModel,
     TradeAccountMatrix,
@@ -169,7 +169,7 @@ def main():
             f"{acc.account_number}: positions={acc.positions}, cash={tam.cash_matrix[acc.account_number]:.2f}"
         )
     if args.exporter and args.export_path:
-        from core.plugins.loader import load_export_plugin
+        from realloc.plugins.loader import load_export_plugin
 
         plugin = load_export_plugin(args.exporter)
         plugin.export(tam.portfolio_trades, args.export_path)
