+++ src/realloc/scripts/end_to_end_1.py
@@ -1,4 +1,4 @@
-from core import *
+from realloc import *
 
 from typing import List, Optional, Dict
 import math
