+++ src/realloc/selectors.py
@@ -1,6 +1,6 @@
 from typing import List, Optional, Dict
 
-from core.accounts import Account
+from realloc.accounts import Account
 
 
 def select_account_for_buy_trade(
