+++ tests/test_matrix.py
@@ -1,5 +1,5 @@
 import pytest
-from core import (
+from realloc import (
     Account,
     TradeAccountMatrix,
     allocate_trades,
